+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/child_metering_points/wrapper.py
@@ -1,16 +1,18 @@
 from pyspark.sql import DataFrame
 from pyspark_functions.data_frame_wrapper import DataFrameWrapper
 
-from opengeh_electrical_heating.infrastructure.electrical_heating_internal.schemas.calculations import calculations
+from opengeh_electrical_heating.infrastructure.electricity_market.child_metering_points.child_metering_points_v1 import (
+    child_metering_points_v1,
+)
 
 
-class Calculations(DataFrameWrapper):
-    """Represents the calculations data structure."""
+class ChildMeteringPoints(DataFrameWrapper):
+    """Represents the child metering points data structure."""
 
     def __init__(self, df: DataFrame):
         super().__init__(
             df,
-            calculations,
+            child_metering_points_v1,
             # We ignore_nullability because it has turned out to be too hard and even possibly
             # introducing more errors than solving in order to stay in exact sync with the
             # logically correct schema.
