+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/consumption_metering_point_periods/schema.py
