+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/repository.py
@@ -2,15 +2,17 @@ from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql import functions as F
 from pyspark.sql import types as T
 
-from opengeh_electrical_heating.infrastructure.electricity_market.data_structure import (
-    ChildMeteringPoints,
-    ConsumptionMeteringPointPeriods,
-)
-from opengeh_electrical_heating.infrastructure.electricity_market.schemas.child_metering_points_v1 import (
+from opengeh_electrical_heating.infrastructure.electricity_market.child_metering_points.schema import (
     child_metering_points_v1,
 )
-from opengeh_electrical_heating.infrastructure.electricity_market.schemas.consumption_metering_point_periods_v1 import (
+from opengeh_electrical_heating.infrastructure.electricity_market.child_metering_points.wrapper import (
+    ChildMeteringPoints,
+)
+from opengeh_electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.schema import (
     consumption_metering_point_periods_v1,
+)
+from opengeh_electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.wrapper import (
+    ConsumptionMeteringPointPeriods,
 )
