+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_bronze/schema.py
