+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_gold/schema.py
