+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements/measurements_gold/wrapper.py
@@ -1,7 +1,7 @@
 from pyspark.sql import DataFrame
 from pyspark_functions.data_frame_wrapper import DataFrameWrapper
 
-from opengeh_electrical_heating.infrastructure.measurements_gold.schemas.time_series_points_v1 import (
+from opengeh_electrical_heating.infrastructure.measurements.measurements_gold.schema import (
     time_series_points_v1,
 )
