--- source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_gold/__init__.py
@@ -1,3 +1,0 @@
-from .repository import Repository
-
-__all__ = ["Repository"]
