+++ source/electrical_heating/tests/electrical_heating_tests/infrastructure/measurements_bronze/test_repository.py
@@ -1,7 +1,7 @@
 from pyspark.sql import SparkSession
 
-from opengeh_electrical_heating.infrastructure.measurements_bronze.data_structure import MeasurementsBronze
-from opengeh_electrical_heating.infrastructure.measurements_bronze.repository import Repository
+from opengeh_electrical_heating.infrastructure.measurements.measurements_bronze.wrapper import MeasurementsBronze
+from opengeh_electrical_heating.infrastructure.measurements.repository import Repository
 
 
 def test__write_measurements__can_be_read(
@@ -13,7 +13,7 @@ def test__write_measurements__can_be_read(
     expected_count = measurements_bronze.df.count()
 
     # Act
-    repository.write_measurements(measurements_bronze, write_mode="overwrite")
+    repository.write_measurements_bronze(measurements_bronze, write_mode="overwrite")
 
     # Assert
-    assert repository.read_measurements().df.count() == excepted_count
+    assert repository.read_measurements_bronze().df.count() == expected_count
