+++ source/electrical_heating/tests/scenario_tests/conftest.py
@@ -10,13 +10,13 @@ from testcommon.etl import TestCase, TestCases
 from opengeh_electrical_heating.application.execute_with_deps import (
     execute_calculation,
 )
-from opengeh_electrical_heating.infrastructure.electricity_market.schemas.child_metering_points_v1 import (
+from opengeh_electrical_heating.infrastructure.electricity_market.child_metering_points.schema import (
     child_metering_points_v1,
 )
-from opengeh_electrical_heating.infrastructure.electricity_market.schemas.consumption_metering_point_periods_v1 import (
+from opengeh_electrical_heating.infrastructure.electricity_market.consumption_metering_point_periods.schema import (
     consumption_metering_point_periods_v1,
 )
-from opengeh_electrical_heating.infrastructure.measurements_gold.schemas.time_series_points_v1 import (
+from opengeh_electrical_heating.infrastructure.measurements.measurements_gold.schema import (
     time_series_points_v1,
 )
 from tests.scenario_tests.electrical_heating_test_args import ElectricalHeatingTestArgs
